Graph design
The basic layout of the chart has been changed to a better contrasting version with larger text and frames around each column for better and easier readability. The chart works fully with the values and even when reopened after shutting down the application, it loads immediately. Occasionally it bugs the appearance to default settings, but probably a bug of android studio.